Moved login to auth stack
