+++ mobile/app/(auth)/login.tsx
@@ -1,35 +1,95 @@
-import { router } from "expo-router";
-import { View, Text, Button, SafeAreaView, StyleSheet } from "react-native";
+import React, { useState } from "react";
+import { Alert, StyleSheet, View } from "react-native";
+import { supabase } from "@/lib/supabase";
+import { Button, Input } from "@rneui/themed";
+import { useRouter } from "expo-router";
 
-export default function Login() {
+export default function Auth() {
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const [loading, setLoading] = useState(false);
+  const router = useRouter();
+
+  async function signInWithEmail() {
+    setLoading(true);
+    const { error } = await supabase.auth.signInWithPassword({
+      email: email,
+      password: password,
+    });
+
+    if (error) Alert.alert(error.message);
+    setLoading(false);
+
+    router.push("/(tabs)");
+  }
+
+  async function signUpWithEmail() {
+    setLoading(true);
+    const { error } = await supabase.auth.signUp({
+      email: email,
+      password: password,
+    });
+
+    if (error) Alert.alert(error.message);
+    else {
+      Alert.alert("Check your email for verification!");
+      router.push("/(tabs)");
+    }
+    setLoading(false);
+  }
+
   return (
-    <SafeAreaView>
-      <View style={styles.textContainer}>
-        <Text style={styles.title}>Login Screen</Text>
+    <View style={styles.container}>
+      <View style={[styles.verticallySpaced, styles.mt20]}>
+        <Input
+          label="Email"
+          leftIcon={{ type: "font-awesome", name: "envelope" }}
+          onChangeText={(text) => setEmail(text)}
+          value={email}
+          placeholder="email@address.com"
+          autoCapitalize={"none"}
+        />
       </View>
-      <Button
-        title="Go to Main App"
-        onPress={() => {
-          router.replace("/(tabs)");
-        }}
-      />
-      <Button
-        title="Go to Sign up"
-        onPress={() => {
-          router.replace("/signup");
-        }}
-      />
-    </SafeAreaView>
+      <View style={styles.verticallySpaced}>
+        <Input
+          label="Password"
+          leftIcon={{ type: "font-awesome", name: "lock" }}
+          onChangeText={(text) => setPassword(text)}
+          value={password}
+          secureTextEntry={true}
+          placeholder="Password"
+          autoCapitalize={"none"}
+        />
+      </View>
+      <View style={[styles.verticallySpaced, styles.mt20]}>
+        <Button
+          title="Sign in"
+          disabled={loading}
+          onPress={() => signInWithEmail()}
+        />
+      </View>
+      <View style={styles.verticallySpaced}>
+        <Button
+          title="Sign up"
+          disabled={loading}
+          onPress={() => signUpWithEmail()}
+        />
+      </View>
+    </View>
   );
 }
 
 const styles = StyleSheet.create({
-  textContainer: {
-    justifyContent: "center",
-    alignItems: "center",
+  container: {
+    marginTop: 40,
+    padding: 12,
   },
-  title: {
-    fontSize: 24,
-    fontWeight: "bold",
+  verticallySpaced: {
+    paddingTop: 4,
+    paddingBottom: 4,
+    alignSelf: "stretch",
+  },
+  mt20: {
+    marginTop: 20,
   },
 });
